Testes Insert Realizados

+++ src/main/java/com/liberandum/Entities/GeoCoord.java
@@ -1,27 +1,27 @@
 package com.liberandum.Entities;
 
 public class GeoCoord {
-    private float latitude = 0;
-    private float longitude = 0;
+    private double latitude = 0;
+    private double longitude = 0;
 
-    public GeoCoord(float latitude, float longitude) {
+    public GeoCoord(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
